Extract duplicate HTTPClientSpy to shared scope
